refactor(config): 字段变更通知直接内联到 config model，用 type(self).model_fields 兼容 Pydantic V3
