+++ src/danmaku_sender/config/sender_config.py
@@ -1,11 +1,31 @@
+from typing import Callable
+
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 
-from .field_watcher import FieldWatcherMixin
 
-
-class SenderConfig(FieldWatcherMixin, BaseModel):
+class SenderConfig(BaseModel):
     """发送器的配置数据"""
     model_config = ConfigDict(validate_assignment=True)
+
+    # 字段变更通知
+    _watchers: dict[str, list[Callable]] = {}
+
+    def __setattr__(self, name: str, value) -> None:
+        super().__setattr__(name, value)
+        if name in type(self).model_fields:
+            for cb in self._watchers.get(name, []):
+                cb(value)
+
+    def watch(self, field: str, callback: Callable) -> None:
+        """注册字段变更回调"""
+        self._watchers.setdefault(field, []).append(callback)
+
+    def unwatch(self, field: str, callback: Callable) -> None:
+        """注销字段变更回调"""
+        try:
+            self._watchers[field].remove(callback)
+        except (KeyError, ValueError):
+            pass
 
     # 延迟设置
     min_delay: float = Field(default=8.0, ge=0.1)
